requirements.txt
gdown
cython
pycocotools
numpy
pandas
Pillow